Rename project from ExpenseOwl to PesaDigital and localize for developing countries
- Rename all occurrences of 'expenseowl' to 'pesadigital' in code, configs, and docs
- Update branding, logo, and project description to focus on financial tracking for developing countries
- Add Tanzanian Shilling (TSH) currency support with proper formatting
- Simplify README to emphasize accessibility and remove complex features
- Remove outdated screenshots and update remaining assets
- Update Docker and Kubernetes configurations with new naming
- Adjust default currency to TSH in the frontend
- Maintain all core functionality while rebranding
