+++ internal/web/templates/manifest.json
@@ -1,6 +1,6 @@
 {
-    "name": "ExpenseOwl",
-    "short_name": "ExpenseOwl",
+    "name": "PesaDigital",
+    "short_name": "PesaDigital",
     "start_url": "/",
     "display": "standalone",
     "background_color": "#1a1a1a",
