+++ internal/web/templates/settings.html
@@ -15,14 +15,14 @@
             }
         })();
     </script>
-    <title>ExpenseOwl Settings</title>
+    <title>PesaDigital Settings</title>
 </head>
 <body>
     <div class="container">
         <header>
             <div class="nav-bar">
                 <a href="/">
-                    <img src="/pwa/icon-192.png" alt="ExpenseOwl Logo" height="85" style="vertical-align: middle; margin-right: 20px;">
+                    <img src="/pwa/icon-192.png" alt="PesaDigital Logo" height="85" style="vertical-align: middle; margin-right: 20px;">
                 </a>
                 <a href="/" class="view-button" data-tooltip="Dashboard">
                     <i class="fa-solid fa-chart-pie"></i>
@@ -37,11 +37,11 @@
         </header>
 
         <div class="info-field">
-            Version: <a href="https://github.com/tanq16/expenseowl/releases/latest" target="_blank" rel="noopener noreferrer">v4.0</a>
+            Version: <a href="https://github.com/isaka-james/pesadigital/releases/latest" target="_blank" rel="noopener noreferrer">v4.0</a>
             <span class="separator">|</span>
-            <a href="https://github.com/tanq16/expenseowl/blob/main/README.md" target="_blank" rel="noopener noreferrer">Documentation</a>
+            <a href="https://github.com/isaka-james/pesadigital/blob/main/README.md" target="_blank" rel="noopener noreferrer">Documentation</a>
             <span class="separator">|</span>
-            <a href="https://github.com/tanq16/expenseowl" target="_blank" rel="noopener noreferrer">GitHub</a>
+            <a href="https://github.com/isaka-james/pesadigital" target="_blank" rel="noopener noreferrer">GitHub</a>
         </div>
 
         <div class="form-container">
@@ -102,7 +102,7 @@
                         <input type="file" id="csv-import-file" accept=".csv" style="display: none;">
                     </div>
                     <div class="import-option">
-                        <label for="csv-import-file-old" class="nav-button">Import from ExpenseOwl v3.20-</label>
+                        <label for="csv-import-file-old" class="nav-button">Import from PesaDigital v3.20-</label>
                         <input type="file" id="csv-import-file-old" accept=".csv" style="display: none;">
                     </div>
                 </div>
@@ -799,7 +799,7 @@
                 fetch('/version').then(r => r.ok ? r.text() : 'dev').then(version => {
                     const link = document.getElementById('version-link');
                     link.textContent = version;
-                    if (version !== 'dev') link.href = `https://github.com/tanq16/expenseowl/releases/tag/${version}`;
+                    if (version !== 'dev') link.href = `https://github.com/isaka-james/pesadigital/releases/tag/${version}`;
                 }).catch(e => console.error('Failed to fetch version:', e));
 
             } catch (error) {
