+++ internal/web/templates/table.html
@@ -15,14 +15,14 @@
             }
         })();
     </script>
-    <title>ExpenseOwl Table</title>
+    <title>PesaDigital Table</title>
 </head>
 <body>
     <div class="container">
         <header>
             <div class="nav-bar">
                 <a href="/">
-                    <img src="/pwa/icon-192.png" alt="ExpenseOwl Logo" height="85" style="vertical-align: middle; margin-right: 20px;">
+                    <img src="/pwa/icon-192.png" alt="PesaDigital Logo" height="85" style="vertical-align: middle; margin-right: 20px;">
                 </a>
                 <a href="/" class="view-button" data-tooltip="Dashboard">
                     <i class="fa-solid fa-chart-pie"></i>
