+++ kubernetes/Expenseowl-configmap.yml
@@ -1,8 +1,8 @@
 apiVersion: v1
 kind: ConfigMap
 metadata:
-  name: expenseowl-config
-  namespace: expenseowl
+  name: pesadigital-config
+  namespace: pesadigital
 data:
   EXPENSE_CATEGORIES: "Food,Groceries,Travel,Rent,Utilities,Money Transfer,Entertainment,Healthcare,Shopping,Other"
   CURRENCY: jpy
