+++ kubernetes/Expenseowl-pvc.yml
@@ -1,8 +1,8 @@
 apiVersion: v1
 kind: PersistentVolumeClaim
 metadata:
-  name: expenseowl-pvc
-  namespace: expenseowl
+  name: pesadigital-pvc
+  namespace: pesadigital
 spec:
   accessModes:
   - ReadWriteOnce
