+++ kubernetes/README.md
@@ -4,14 +4,14 @@ Use the following instructions to apply the Kubernetes spec:
 
 ```bash
 kubectl apply -f kubernetes/_namespace.yml
-kubectl apply -f kubernetes/Expenseowl-Deployment.yml
-kubectl apply -f kubernetes/Expenseowl-configmap.yml
-kubectl apply -f kubernetes/Expenseowl-svc.yml
-kubectl apply -f kubernetes/Expenseowl-pvc.yml
-kubectl apply -f kubernetes/Expenseowl-ingress.yml
+kubectl apply -f kubernetes/Pesadigital-Deployment.yml
+kubectl apply -f kubernetes/Pesadigital-configmap.yml
+kubectl apply -f kubernetes/Pesadigital-svc.yml
+kubectl apply -f kubernetes/Pesadigital-pvc.yml
+kubectl apply -f kubernetes/Pesadigital-ingress.yml
 kubectl port-forward pod/<pod-name> 8080:8080 # Change Pod Name Here
 ```
 
 ```
-Dashboard available at http://expenseowl.localhost/
+Dashboard available at http://pesadigital.localhost/
 ```
