+++ kubernetes/_namespace.yml
@@ -1,4 +1,4 @@
 apiVersion: v1
 kind: Namespace
 metadata:
-  name: expenseowl
+  name: pesadigital
